% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_dist.R
\name{AbsCorDist}
\alias{AbsCorDist}
\title{AbsCorDist}
\usage{
AbsCorDist(dataset, nthreads = NULL)
}
\arguments{
\item{dataset}{A dataset created my using \link{MedList} on a transcriptomics
dataset}

\item{nthreads}{Number of processor threads for the filtering. If not
specifed then the maximum number of logical cores are used.}
}
\description{
Calculates a distance matrix based on the distance measure of:
\deqn{1 - |cor(x, y)|}
}
\examples{
meds <- MedList(Laurasmappings, nthreads = 2)
distance <- AbsCorDist(meds, nthreads = 2)
}
