% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{ActivitySelect}
\alias{ActivitySelect}
\title{ActivitySelect}
\usage{
ActivitySelect(ID, dataset)
}
\arguments{
\item{ID}{Gene name or row number}

\item{dataset}{A transcriptomics dataset. First columns should be gene names. All other columns should be expression levels.}
}
\value{
Gene activity as a matrix
}
\description{
Returns gene activity by either gene name or row number
}
\examples{
activity <- ActivitySelect(1, Laurasmappings)
activity <- ActivitySelect("comp100000_c0_seq2", Laurasmappings)
}
