% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovafilter.R
\name{anovafilter}
\alias{anovafilter}
\title{anovafilter:
filters a genedata via}
\usage{
anovafilter(dataset)
}
\arguments{
\item{dataset}{A transcriptomics dataset. First columns should be gene names. All other columns should be expression levels.}
}
\description{
anovafilter:
filters a genedata via
}
\examples{
Laurasmappings_filtered <- cosinoranalysis(Laurasmappings)
}
