% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicplot.R
\name{basicplot}
\alias{basicplot}
\title{basicplot:
Plots activity data as points and mean data as lines}
\usage{
basicplot(genename, dataset, timelag = 0, save = FALSE, print = TRUE)
}
\arguments{
\item{genename}{The name of a gene intended for plotting. Must be a string.}

\item{dataset}{A transcriptomics dataset. First columns should be gene names. All other columns should be expression levels.}

\item{timelag}{Shifts the plot to earlier in time.}

\item{save}{Logical. If TRUE, saves plot to working directory. Defaults to FALSE.}

\item{print}{Logical. If TRUE renders plot in the plot viewer. Defaults to TRUE}
}
\value{
Prints or saves ggplot2 object(s)
}
\description{
basicplot:
Plots activity data as points and mean data as lines
}
\examples{
basicplot('comp101252_c0_seq2',LauraSingleMap)
}
