% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircadianTools.R
\docType{package}
\name{CircadianTools}
\alias{CircadianTools}
\alias{CircadianTools-package}
\title{CircadianTools: Collection of Tools for Detecting Rhythmic Genes}
\description{
Provides functions to detect rhythmicity using Cosinor models or turning point analysis.
}
