% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{ClusterCenterGenerator}
\alias{ClusterCenterGenerator}
\title{ClusterCenterGenerator}
\usage{
ClusterCenterGenerator(cluster.dataset, nthreads = NULL)
}
\arguments{
\item{cluster.dataset}{A transcriptomics dataset where the final column
details the cluster the gene belongs to. First column should be gene names.
All remaining columns should be expression levels.}

\item{nthreads}{Number of processor threads for the process. If not specifed
then the maximum number of logical cores are used.}
}
\description{
Finds the center of every cluster in a dataset
}
\examples{
filter.df <- CombiFilter(Laurasmappings)
pam.df <- PamClustering(filter.df, k = 10)
ClusterCenterGenerator(pam.df)
}
