% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{ClusterTimeProfile}
\alias{ClusterTimeProfile}
\title{ClusterTimeProfile}
\usage{
ClusterTimeProfile(cluster.no, cluster.dataset, nthreads = NULL)
}
\arguments{
\item{cluster.no}{The number which identifies the cluster.}

\item{cluster.dataset}{A transcriptomics dataset where the final column
details the cluster the gene belongs to. First column should be gene names.
All remaining columns should be expression levels.}

\item{nthreads}{The number of threads to be used for parallel computations.
Defaults to the maximum number of threads available.}
}
\description{
Provides a vector of mean values at each time point for
 each cluster.
}
\examples{
pam.df <- PamClustering(Laurasmappings, k = 10, nthreads = 1)
time.profile<-ClusterTimeProfile(1, cluster.dataset = pam.df, nthreads = 1)

}
