% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combifilter.R
\name{combifilter}
\alias{combifilter}
\title{combifilter}
\usage{
combifilter(dataset, non_zero_num = 4, threshold = 0.05,
  cutoff = 0.1, nthreads = NULL)
}
\arguments{
\item{dataset}{A transcriptomics dataset. First columns should be gene names. All other columns should be expression levels.}

\item{non_zero_num}{The minimum number of non-zero activity readings for genes in the filtered dataset}

\item{threshold}{Set the p-value threshold for the filtering}

\item{cutoff}{Proportion of total number of genes (which show at least some activity) to be removed}

\item{nthreads}{Number of processor threads for the filtering. If not specifed then the maximum number of logical cores are used.}
}
\description{
Filters a transcriptomics dataset by using zerofilter, anovafilter and sizefilter
}
\examples{
Laurasmappings_filtered <- combifilter(Laurasmappings, nthreads=4)

}
