% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_single_cluster.R
\name{CommonSingletonFinder}
\alias{CommonSingletonFinder}
\title{CommonSingletonFinder}
\usage{
CommonSingletonFinder(cluster.dataset1, cluster.dataset2)
}
\arguments{
\item{cluster.dataset1}{A transcriptomics dataset where the final column details the cluster the gene belongs to. First column should be gene names. All remaining columns should be expression levels.}

\item{cluster.dataset2}{A transcriptomics dataset in the same format as cluster.dataset1 but generated via a different clustering method or with different parameters.}
}
\description{
Finds the genes which belong to common singleton clusters in two different clustered datasets.
}
\examples{
filter.df <- CombiFilter(Laurasmappings)
pam.df <- PamClustering(filter.df, 40)
hclust.df <- HClustering(filter.df, 40)
common.singletons <- CommonSingletonFinder(pam.df)

}
