% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compplot.R
\name{compplot}
\alias{compplot}
\title{compplot:}
\usage{
compplot(gene1, gene2, dataset, save = FALSE)
}
\arguments{
\item{gene1}{The name of the first gene to be plotted. Must be a string.}

\item{gene2}{The name of the second gene to be plotted. Must be a string.}

\item{dataset}{A transcriptomics dataset. First columns should be gene names. All other columns should be expression levels.}

\item{save}{Logical. If TRUE, saves plots to working directory. Defaults to FALSE.}
}
\value{
Returns or saves ggplot2 object(s).
}
\description{
Plots two genes from a gene activity dataset
}
\examples{
cor_results <- coranalysis(LauraSingleMap, 'comp100002_c0_seq2')
}
