% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{ContigGen}
\alias{ContigGen}
\title{ContigGen}
\usage{
ContigGen(dataset)
}
\arguments{
\item{dataset}{A transcriptomics dataset. First columns should be gene names.
All other columns should be expression levels.}
}
\value{
A character vector of contig IDs
}
\description{
Finds all unique contig IDs in a transcriptomics dataset
}
\examples{
contigs <- ContigGen(Laurasmappings)

}
