% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_functions.R
\name{CorAnalysisClusterDataset}
\alias{CorAnalysisClusterDataset}
\title{CorAnalysisClusterDataset:}
\usage{
CorAnalysisClusterDataset(
  cluster.dataset,
  lag = 0,
  nthreads = NULL,
  save = TRUE,
  filename = NULL
)
}
\arguments{
\item{cluster.dataset}{A transcriptomics dataset where the final column
details the cluster the gene belongs to. First column should be gene names.
All remaining columns should be expression levels.}

\item{lag}{Setting any value other than 0 allows a gene to be correlated with
lagged genes in the dataset. The number denotes the number of timesteps to
lag by.}

\item{nthreads}{The number of threads to be used for parallel computations.
Defaults to the maximum number of threads available.}

\item{save}{Logical. If TRUE, the dataframe of correlations for the dataset
is saved as a .csv file.}

\item{filename}{filename for saved csv file. Only used if save=TRUE. If not
specified then the dataset object name is used.}
}
\value{
A dataframe of correlation values. The column genes represent the
 original clusters whilst the rows represent lagged clusters.
}
\description{
Correlates the average activity of each cluster with every other
 cluster in a dataset.
}
\examples{
pam.df <- PamClustering(Laurasmappings, k = 10, nthreads = 2)
cor.df <- CorAnalysisClusterDataset(pam.df, nthreads = 2)

}
