% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coranalysispar.R
\name{coranalysispar}
\alias{coranalysispar}
\title{coranalysispar:}
\usage{
coranalysispar(genename, dataset, nthreads = NULL)
}
\arguments{
\item{genename}{the name of a gene intended for comparison with all other genes in the dataset. Must be a string.}

\item{dataset}{A transcriptomics dataset. First columns should be gene names. All other columns should be expression levels.}

\item{nthreads}{Number of processor threads for the process. If not specifed then the maximum number of logical cores are used.}
}
\value{
Returns dataframe containing gene names and correlation values
}
\description{
Parallel Implementation of \link{coranalysis}. Ranks correlation between a given gene and all over genes in a dataset. Plots both the given gene and highly correlated genes for a given correlation value
}
\examples{
cor_results <- coranalysis('comp100002_c0_seq2',Laurasmappings)
}
