% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corsignificantplot.R
\name{corsignificantplot}
\alias{corsignificantplot}
\title{corsignificantplot :}
\usage{
corsignificantplot(results, dataset, number = 10, print = TRUE,
  save = FALSE, negative = FALSE)
}
\arguments{
\item{results}{A dataframe generated by \code{cosinoranalysis}}

\item{dataset}{A transcriptomics dataset which was used with \code{cosinoranalysis} to generate the results dataframe. First columns should be gene names. All other columns should be expression levels.}

\item{number}{The number of most significant genes printed or saved}

\item{print}{Logical. If TRUE renders significant genes in the plot viewer. Defaults to TRUE}

\item{save}{Logical. If TRUE, saves plots to working directory. Defaults to FALSE.}

\item{period}{The period of rhythmicity which is being tested for. Defaults to 24 (circadian).}
}
\value{
Prints or saves ggplot2 object(s).
}
\description{
Prints or saves the genes found to be most significant by \link{cosinoranalysis}
}
\examples{
cor_results <- coranalysis('comp100002_c0_seq2',Laurasmappings)
corsignificantplot(cor_results, Laurasmappings, number = 15,save=TRUE, negative=FALSE)
}
