% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinor_functions.R
\name{CosinorAnalysisPar}
\alias{CosinorAnalysisPar}
\title{CosinorAnalysisPar:}
\usage{
CosinorAnalysisPar(dataset, period = 24, nthreads = NULL, timelag = 6)
}
\arguments{
\item{dataset}{A transcriptomics dataset. First columns should be gene names.
All other columns should be expression levels.}

\item{period}{The period of rhythmicity which is being tested for. Defaults
to 24 (circadian).}

\item{nthreads}{The number of threads to be used for parallel computations.
Defaults to the maximum number of threads available.}

\item{timelag}{The value of time for which observations begin. Defaults to 6.}
}
\value{
Saves ggplot2 object(s).
}
\description{
Parallel Implementation of \link{CosinorAnalysis}. Fits cosinor
 models to transcriptomics data and plots the best-fitting models using
 ggplot2.
}
\examples{
cosinor.results <- CosinorAnalysisPar(Laurasmappings, nthreads = 2)

}
