% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinorplot.R
\name{cosinorplot}
\alias{cosinorplot}
\title{cosinorplot:
Fits a cosinor model to a given gene in a given dataset and plots the model}
\usage{
cosinorplot(genename, dataset, timelag = 6, period = 24,
  print = TRUE, save = FALSE)
}
\arguments{
\item{genename}{The name of a gene intended for plotting. Must be a string.}

\item{dataset}{A transcriptomics dataset. First columns should be gene names. All other columns should be expression levels.}

\item{timelag}{The value of time for which observations begin. Defaults to 6.}

\item{print}{Logical. If TRUE renders plot in the plot viewer. Defaults to TRUE}

\item{save}{Logical. If TRUE, saves plot to working directory. Defaults to FALSE.}
}
\value{
Prints or saves ggplot2 object(s)
}
\description{
cosinorplot:
Fits a cosinor model to a given gene in a given dataset and plots the model
}
\examples{
cosinorplot('comp100000_c0_seq2', LauraSingleMap )
}
