% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinor_residuals.R
\name{CosinorResidualPlot}
\alias{CosinorResidualPlot}
\title{CosinorResidualPlot}
\usage{
CosinorResidualPlot(genename, dataset, timelag = 6, period = 24,
  print = TRUE, save = FALSE, path = NULL)
}
\arguments{
\item{genename}{The name of a gene intended for plotting. Must be a string.}

\item{dataset}{A transcriptomics dataset. First columns should be gene names.
All other columns should be expression levels.}

\item{timelag}{Shifts the plot to earlier in time.}

\item{period}{The period of rhythmicity which is being tested for.
Defaults to 24 (circadian).}

\item{print}{Logical. If TRUE renders plot in the plot viewer. Defaults to
TRUE}

\item{save}{Logical. If TRUE, saves plot to working directory. Defaults to
FALSE.}

\item{path}{The directory to be used for saving plots to. Uses the working
directory by default. Not used if save=FALSE}
}
\value{
Prints or saves ggplot2 object(s)
}
\description{
Fits a cosinor model to a gene and plot the residuals
}
\examples{
CosinorResidualPlot('comp99801_c1_seq1', Laurasmappings)

}
