% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinorsignificantplot.R
\name{cosinorsignificantplot}
\alias{cosinorsignificantplot}
\title{cosinorsignificantplot :
Prints or saves the genes found to be most significant by \code{cosinoranalysis}}
\usage{
cosinorsignificantplot(results, dataset, number = 10, period = 24,
  print = TRUE, save = FALSE)
}
\arguments{
\item{results}{A dataframe generated by \code{cosinoranalysis}}

\item{dataset}{A transcriptomics dataset which was used with \code{cosinoranalysis} to generate the results dataframe. First columns should be gene names. All other columns should be expression levels.}

\item{number}{The number of most significant genes printed or saved}

\item{period}{The period of rhythmicity which is being tested for. Defaults to 24 (circadian).}

\item{print}{Logical. If TRUE renders significant genes in the plot viewer. Defaults to TRUE}

\item{save}{Logical. If TRUE, saves plots to working directory. Defaults to FALSE.}
}
\value{
Prints or saves ggplot2 object(s).
}
\description{
cosinorsignificantplot :
Prints or saves the genes found to be most significant by \code{cosinoranalysis}
}
\examples{
cosinor_results <- cosinoranalysis(Laurasmappings)
cosinorsignificantplot(cosinor_results, Laurasmappings, number = 15, period=24 ,save=TRUE)
}
