% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoscape_functions.R
\name{CytoscapeFilter}
\alias{CytoscapeFilter}
\title{CytoscapeFilter}
\usage{
CytoscapeFilter(cytoscape.file, threshold = 0.9, save = TRUE, filename = NULL)
}
\arguments{
\item{cytoscape.file}{A dataframe object generated by CytoscapeFile}

\item{threshold}{Positive genes below this threshold and negative genes above
the additive inverse of this threshold will be removed.}

\item{save}{Logical. If TRUE then the dataframe will be saved as a csv file.}

\item{filename}{String. The name for the saved file.}
}
\description{
Reduces the size of a file intended for Cytoscape by filtering
 out the genes/clusters which are not correlated
}
\examples{
filter.df <- CombiFilter(Laurasmappings, nthreads = 2)
pam.df <- PamClustering(filter.df, k = 10, nthreads = 2)
cor.df <- CorAnalysisClusterDataset(pam.df, save=FALSE, nthreads = 2)
cyto.df <- CytoscapeFile(cor.df, save = FALSE, nthreads = 2)
cyto.filtered.df <- CytoscapeFilter(cyto.df, threshold=0.95)

}
