% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DendogramPlot.R
\name{DatasetDendogram}
\alias{DatasetDendogram}
\title{DatasetDendogram}
\usage{
DatasetDendogram(
  cluster.dataset,
  method = "agglom",
  metric = "euclidean",
  nthreads = NULL,
  print = TRUE,
  save = FALSE,
  path = NULL
)
}
\arguments{
\item{cluster.dataset}{A transcriptomics dataset where the final column
details the cluster the gene belongs to. First column should be gene names.
All remaining columns should be expression levels.}

\item{method}{The clustering method to be used. Currently accepts 'agglom'
and 'diana'.}

\item{metric}{The distance metric to be used to calculate the distances
between genes. See parallelDist::parDist for all accepted arguments. Also
allows the option of 'abs.correlation'.}

\item{nthreads}{Number of processor threads to be used for calculating the
distance matrix. If not specified then the maximum number of logical cores
is used.}

\item{print}{Logical. If TRUE renders plot in the plot viewer. Defaults to
TRUE}

\item{save}{Logical. If TRUE, saves plot to working directory. Defaults to
FALSE.}

\item{path}{The directory to be used for saving plots to. Uses the working
directory by default. Not used if save = FALSE}
}
\description{
DatasetDendogram
}
\examples{
diana.df <- DianaClustering(Laurasmappings, k = 10, nthreads = 2)
DatasetDendogram(diana.df, method = 'diana', nthreads = 2, path = 'example')
}
