% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DendogramPlot.R
\name{DendogramPlot}
\alias{DendogramPlot}
\title{DendogramPlot}
\usage{
DendogramPlot(cluster.no, cluster.dataset, method = "agglom",
  metric = "euclidean", print = TRUE, save = FALSE, path = NULL)
}
\arguments{
\item{cluster.no}{The number which identifies the cluster. If a vector is
provided then multiple clusters will be used.}

\item{cluster.dataset}{A transcriptomics dataset where the final column
details the cluster the gene belongs to. First column should be gene names.
All remaining columns should be expression levels.}

\item{method}{The clustering method to be used. Currently accepts 'agglom'
and 'diana'.}

\item{metric}{The distance metric to be used to calculate the distances
between genes. See parallelDist::parDist for all accepted arguments. Also
allows the option of 'abs.correlation'.}

\item{print}{Logical. If TRUE renders plot in the plot viewer. Defaults to
TRUE}

\item{save}{Logical. If TRUE, saves plot to working directory. Defaults to
FALSE.}

\item{path}{The directory to be used for saving plots to. Uses the working
directory by default. Not used if save = FALSE}
}
\description{
Plots the dendogram for a cluster in a clustered dataset.
}
\examples{
t.filter <-TFilter(Laurasmappings)
diana.df <- DianaClustering(t.filter, k= 10)
DendogramPlot(5, diana.df, method = 'diana')
}
