% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diana.R
\name{DianaParamSelection}
\alias{DianaParamSelection}
\title{DianaParamSelection}
\usage{
DianaParamSelection(dataset, distance = NULL, k = c(2, 5, 10),
  metric = "euclidean", nthreads = 4)
}
\arguments{
\item{dataset}{A transcriptomics dataset. Preferably filtered first. First columns should be gene names. All other columns should be expression levels.}

\item{distance}{A distance matrix. If a distance matrix has already been created (such as by using the DistanceGen function), the matrix can be passed to this function to save time. If a distance matrix is not provided then it will be generated by the function.}

\item{k}{A numeric vector giving the number of clusters to be evaluated.}

\item{metric}{The distance metric to be used to calculate the distances between genes. See parallelDist::parDist for all accepted arguments. Also allows the option of "abs.correlation". Not used if a distance matrix is provided.
filter.df <- CombiFilter(Laurasmappings)
k.options <- seq(10,100, by=10)
diana.validation <- DianaParamSelection(filterdf, k=k.options)}

\item{nthreads}{The number of threads to be used for parallel computations.If NULL then the maximum number of threads available will be used.}
}
\description{
Runs DIANA (DIvisive ANAlysis) clustering with differing numbers of partitions and returns validation metrics.
}
