% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_gen.R
\name{DistanceGen}
\alias{DistanceGen}
\title{DistanceGen}
\usage{
DistanceGen(dataset, metric = "euclidean", nthreads = NULL)
}
\arguments{
\item{dataset}{A transcriptomics dataset. Preferably filtered first. First
columns should be gene names. All other columns should be expression levels.}

\item{metric}{The distance metric to be used to calculate the distances
between genes. See parallelDist::parDist for all accepted arguments. Also
allows the option of 'abs.correlation'. Not used if a distance matrix is
provided.}

\item{nthreads}{The number of threads to be used for parallel computations.
If NULL then the maximum number of threads available will be used.}
}
\description{
Generates a distance matrix from a a transcriptomics dataset.
}
\examples{
a.filter <- AnovaFilter(Laurasmappings)
distance <- DistanceGen(a.filter, metric='abs.correlation')

}
