% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{FastaSub}
\alias{FastaSub}
\title{FastaSub}
\usage{
FastaSub(gene.names, fasta.file, save = TRUE, filename = NULL)
}
\arguments{
\item{gene.names}{A character vector of gene names}

\item{fasta.file}{A fasta file read in using \code{seqinr}}

\item{save}{Logical. If TRUE then the newly created fasta file is saved to
the working directory}

\item{filename}{The filename to use for a saved fasta file}
}
\value{
A fasta file
}
\description{
Creates a fasta file from only certain sequences in another
 fasta file
}
\examples{
main.fasta <- seqinr::read.fasta("~/JoesTranscriptomeMin300bp2.fasta")
a.filter <- AnovaFilter(Laurasmappings)
fasta.sub <- FastaSub(a.filter$sample, main.fasta, filename="a_filter")
}
