% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{FileConflict}
\alias{FileConflict}
\title{FileConflict}
\usage{
FileConflict(filename)
}
\arguments{
\item{filename}{The name of a file (with extension included)}
}
\description{
Checks if a file which will be created already exists and, if
 necessary asks the user if this file should be overwritten.
}
\examples{
FileConflict('correlation.csv')
}
