% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_distances.R
\name{FindClusterDistanceQuantiles}
\alias{FindClusterDistanceQuantiles}
\title{FindClusterDistanceQuantiles}
\usage{
FindClusterDistanceQuantiles(cluster.dataset, nthreads = NULL)
}
\arguments{
\item{cluster.dataset}{A transcriptomics dataset where the final column
details the cluster the gene belongs to. First column should be gene names.
All remaining columns should be expression levels.}

\item{nthreads}{Number of processor threads for the process. If not specifed
then the maximum number of logical cores are used.}
}
\description{
Finds The distances between the center of each cluster and the
centers of all the other clusters.
}
\examples{
filter.df <- CombiFilter(Laurasmappings)
pam.df <- PamClustering(filter.df, k = 75)
FindClusterDistanceQuantiles(pam.df)
}
