% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescale.R
\name{genescale}
\alias{genescale}
\title{#' genescale:}
\usage{
genescale(dataset, scale = FALSE, center = TRUE)
}
\arguments{
\item{dataset}{A transcriptomics dataset. First columns should be gene names. All other columns should be expression levels.}
}
\value{
A transcriptomics dataset with centered/scaled genes
}
\description{
Centers/scales every gene in a transcriptomics dataset
}
\examples{
geneclean(LauraSingleMap)
}
