% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Laurasmappings}
\alias{Laurasmappings}
\title{Gene expression count data of \emph{T. saltator}.}
\format{A data frame with 91311 rows and 32 variables:
\describe{
  \item{sample}{Sample ID}
  \item{CT6.2}{Gene expression count at t = 6, replicate 1}
  \item{CT6.3}{Gene expression count at t = 6, replicate 2}
  ...
  \item{CT27.5}{Gene expression count at t = 27, replicate 4}
}}
\source{
\url{http://doi.org/10.7717/peerj.2555}
}
\usage{
Laurasmappings
}
\description{
A transcriptomics dataset containing the gene expression counts for 156767
genes of \emph{T. saltator} across 21 hours. First reading is at hour 6.
Each time point has 4 replicates with the exception of hour 18 which has 3.
}
\keyword{datasets}
