% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maketimevector.r
\name{maketimevector}
\alias{maketimevector}
\title{maketimevector:}
\usage{
maketimevector(dataset)
}
\arguments{
\item{dataset}{A transcriptomics dataset. First columns should be gene names. All other columns should be expression levels.}
}
\value{
A vector of time values for the genes
}
\description{
Produces a vector of time values for the gene activity readings.
}
\examples{
maketimevector(LauraSingleMap)
}
