% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdplot.R
\name{MDSPlot}
\alias{MDSPlot}
\title{MDSPlot}
\usage{
MDSPlot(cluster.dataset, title = "Multidimensional Scaling Plot",
  nthreads = NULL, metric = "euclidean", save = FALSE,
  print = TRUE)
}
\arguments{
\item{cluster.dataset}{A transcriptomics dataset where the final column
details the cluster the gene belongs to. First column should be gene names.
All remaining columns should be expression levels.}

\item{title}{The title to be used in the plot}

\item{nthreads}{Number of processor threads for the process. If not
specified then the maximum number of logical cores are used.}

\item{metric}{The distance metric to be used to calculate the distances
between genes. See parallelDist::parDist for all accepted arguments. Also 
accepts "abs.correlation" for absolute Pearson's correlation}

\item{save}{Logical. If TRUE, saves plots. Defaults to FALSE.}

\item{print}{Logical. If TRUE renders significant genes in the plot viewer.
Defaults to TRUE}
}
\value{

}
\description{
Applies multidimensional scaling to a clustered transcriptomics
 dataset to reduce the clusters to two dimensions and then plots the clusters.
}
\examples{
a.filter <- AnovaFilter(Laurasmappings)
cluster.df <- PamClustering(a.filter, k = 95, scale = TRUE)
MDSPlot(cluster.df)
}
