% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_cosinor.R
\name{MultiCosinorTest}
\alias{MultiCosinorTest}
\title{MultiCosinorTest}
\usage{
MultiCosinorTest(
  genename,
  dataset,
  timelag = 6,
  period.1 = 24,
  period.2 = 12.4,
  save = TRUE,
  path = NULL
)
}
\arguments{
\item{genename}{The name of a gene intended for plotting. Must be a string.}

\item{dataset}{A transcriptomics dataset. First columns should be gene names.
All other columns should be expression levels.}

\item{timelag}{The value of time for which observations begin. Defaults to 6.}

\item{period.1}{The period of the simple Cosinor Model.}

\item{period.2}{The period of the additional cosinor terms in the more
complex model.}

\item{save}{Logical. If TRUE then the ANOVA test results will be saved to
three csv files.}

\item{path}{The directory to be used for saving ANOVA results to. Only used
if save == TRUE Uses theprovided genename appended with '_cosinor_tests' if
this argument is not specified.}
}
\description{
Fits a cosinor model and carries out ANOVA using raw
 coefficients. Then fits a cosinor model with additonal sine and cosine terms
 with a different period. ANOVA tests are carried out on the more complex
 model as well as directly comparing the two models.
}
\examples{
MultiCosinorTest("comp100000_c0_seq2", Laurasmappings)
}
