% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{PamClustering}
\alias{PamClustering}
\title{PamClustering:}
\usage{
PamClustering(dataset = NULL, distance = NULL, k,
  metric = "euclidean", nthreads = NULL, scale = FALSE,
  center = TRUE)
}
\arguments{
\item{dataset}{A transcriptomics dataset. First columns should be gene names.
All other columns should be expression levels. Not needed if an argument to
distance is given.}

\item{distance}{A distance matrix. If a distance matrix has already been
created (such as by using the \code{\link{DistanceGen}} function), the
matrix can be passed to this function to save time.
If a distance matrix is not provided then it will be generated by the
function.}

\item{k}{The total number of clusters.}

\item{metric}{The distance metric to be used to calculate the distances
between genes. See \code{parallelDist::parDist} for all accepted arguments.}

\item{nthreads}{Number of processor threads to be used for calculating the
distance matrix. If not specified then the maximum number of logical cores
 are used.}

\item{scale}{If the gene activity should be scaled before clustering.}

\item{center}{If the gene activity should be centered before clustering.}
}
\value{
Returns a transcriptomics dataset provided with an additional cluster
 column appended which denotes which cluster each gene belongs to.
}
\description{
Applies PAM (Partitioning around Medoids) clustering to a
 transcriptomics dataset and appends a cluster column to this dataset for
 each genes.
}
\examples{
filterdf <-TFilter(Laurasmappings)
pam.df <- PamClustering(filterdf,k = 20)
}
