% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_param_selection.R
\name{PamParamSelection}
\alias{PamParamSelection}
\title{PamParamSelection}
\usage{
PamParamSelection(dataset, k = c(2, 5, 10), metric = "euclidean",
  nthreads = 2)
}
\arguments{
\item{dataset}{A transcriptomics dataset. Preferably filtered first. First columns should be gene names. All other columns should be expression levels.}

\item{k}{A numeric vector giving the number of clusters to be evaluated.}

\item{metric}{The distance metric to be used to calculate the distances between genes. See parallelDist::parDist for all accepted arguments. Also allows the option of "abs.correlation".}

\item{nthreads}{The number of threads to be used for parallel computations. Warning! Increasing this value will massively increase RAM usage. If NULL then the maximum number of threads available will be used.}
}
\description{
Runs PAM with differing numbers of partitions and returns validation metrics.
}
\examples{
filter.df <- CombiFilter(Laurasmappings)
k.options <- seq(10,100, by=10)
pam.validation <- PamParamSelection(filterdf, k=k.options)
}
