% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{QuantilePlots}
\alias{QuantilePlots}
\title{QuantilePlots}
\usage{
QuantilePlots(cluster.dataset, metric = "euclidean", nthreads = NULL,
  save = TRUE, print = TRUE, path = NULL)
}
\arguments{
\item{cluster.dataset}{A transcriptomics dataset where the final column
details the cluster the gene belongs to. First column should be gene names.
All remaining columns should be expression levels.}

\item{metric}{The distance metric to be used to calculate the distances
between genes. See parallelDist::parDist for all accepted arguments. Also
allows the option of 'abs.correlation'.}

\item{nthreads}{Number of processor threads for the process. If not specifed
then the maximum number of logical cores are used.}

\item{save}{Logical. If TRUE, saves the histogram plots to working directory.
Defaults to FALSE.}

\item{print}{Logical. If TRUE renders the histogram plots in the plot viewer.
Defaults to TRUE}

\item{path}{The directory to be used for saving plots to. Uses the working
directory by default. Not used if save=FALSE}
}
\description{
Finds the quartiles for intercluster distances and plots these
 distances as a set of histograms
}
\examples{
a.filter <-AnovaFilter(Laurasmappings)
pam.anova <- PamClustering(a.filter, k = 10)
QuantilePlots(pam.anova, path='Pam_Anova_Distance_Histograms')
}
