% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_functions.R
\name{RainAnalysis}
\alias{RainAnalysis}
\title{RainAnalysis :}
\usage{
RainAnalysis(dataset, period)
}
\arguments{
\item{dataset}{A transcriptomics dataset. First columns should be gene names.
All other columns should be expression levels.}
}
\value{
A dataframe object detailing the result of the rain analysis.
}
\description{
Carries out RAIN analysis on a transcriptomics dataset
}
\examples{
results <- RainAnalysis(Laurasmappings, period = 24)
}
