% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{SingletonNameFinder}
\alias{SingletonNameFinder}
\title{SingletonNameFinder}
\usage{
SingletonNameFinder(cluster.dataset)
}
\arguments{
\item{cluster.dataset}{A transcriptomics dataset where the final column details the cluster the gene belongs to. First column should be gene names. All remaining columns should be expression levels.}
}
\value{
Returns a vector of gene names for all genes belonging to singleton clusters in the provided dataset.
}
\description{
Finds the genes which belong to singleton clusters.
}
\examples{
filter.df <- CombiFilter(Laurasmappings)
pam.df <- PamClustering(filter.df, k = 40)
singletons <- SingletonNameFinder(pam.df)


}
