% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizefilter.R
\name{SizeFilter}
\alias{SizeFilter}
\title{SizeFilter}
\usage{
SizeFilter(dataset, cutoff = 0.1, nthreads = NULL)
}
\arguments{
\item{dataset}{A transcriptomics dataset. First columns should be gene names. All other columns should be expression levels.}

\item{cutoff}{Proportion of total number of genes (which show at least some activity) to be removed}

\item{nthreads}{Number of processor threads for the filtering. If not specifed then the maximum number of logical cores are used.}
}
\description{
Filters the genes with the smallest range from a transcriptomics dataset
}
\examples{
rangedf <- SizeFilter(Laurasmappings, nthreads=4)

}
