% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{TAnalysis}
\alias{TAnalysis}
\title{TAnalysis:}
\usage{
TAnalysis(row.no, dataset, psignificance = 0.05)
}
\arguments{
\item{row.no}{The row number of the gene of interest}

\item{dataset}{A transcriptomics dataset. First columns should be gene names. All other columns should be expression levels.}

\item{psignificance}{The maximum p-value for which a result of a t-test is classed as significant}
}
\value{
The number of signficant increases and decreases as a vector
}
\description{
A t-test is carried out on gene activity levels between time points and the number of significant increases & decreases is returned.
}
\examples{
ups.downs <- TAnalysis(row.no = 1, dataset = Laurasmappings,  psignificance = 0.01)

}
