% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{TFilter}
\alias{TFilter}
\title{TFilter:}
\usage{
TFilter(
  dataset,
  maxdifference = 1,
  minchanges = 2,
  psignificance = 0.05,
  nthreads = NULL
)
}
\arguments{
\item{dataset}{A transcriptomics dataset. First columns should be gene names.
All other columns should be expression levels.}

\item{maxdifference}{The maximum difference between the number of signifcant
positive t statistic and the number of signifcant negative t statistic.}

\item{minchanges}{The minimum number of significant changes found via t-tests
for the gene to be included in the filtered dataset.}

\item{psignificance}{The maximum p-value for which a result of a t-test is
classed as significant.}

\item{nthreads}{Number of processor threads to be used for calculating the
distance matrix. If not specified then the maximum number of logical cores
are used.}
}
\value{
Returns a filtered transcriptomics dataset
}
\description{
Applies a filter where a t-test is carried out on
 gene activity levels between time points. The number of significant changes
 between time points is found. If there is a sufficient number of significant
 changes and close to as many positive changes as negative changes then the
 gene is included in the filtered dataset.
}
\examples{
filterdf <- TFilter(Laurasmappings, nthreads = 2)

}
