% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{ZeroFilter}
\alias{ZeroFilter}
\title{ZeroFilter}
\usage{
ZeroFilter(dataset, non_zero_num = 4, nthreads = NULL)
}
\arguments{
\item{dataset}{A transcriptomics dataset. First columns should be gene names. All other columns should be expression levels.}

\item{non_zero_num}{The minimum number of non-zero activity readings for genes in the filtered dataset}

\item{nthreads}{Number of processor threads for the filtering. If not specifed then the maximum number of logical cores are used.}
}
\description{
Filters a transcriptomics dataset such that there is a minimum number of non-zero activity readings for each gene.
}
\examples{
Laurasmappings_filtered <- ZeroFilter(Laurasmappings, nthreads=4)

}
