% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovafilter.R
\name{anovafilter}
\alias{anovafilter}
\title{anovafilter:
filters a genedata via}
\usage{
anovafilter(dataset, threshold = 0.05, nthreads = NULL)
}
\arguments{
\item{dataset}{A transcriptomics dataset. First columns should be gene names. All other columns should be expression levels.}

\item{threshold}{Set the p-value threshold for the filtering}

\item{nthreads}{Number of processor threads for the filtering. If not specifed then the maximum logical cores are used.}
}
\description{
anovafilter:
filters a genedata via
}
\examples{
Laurasmappings_filtered <- anovafilter(Laurasmappings, nthreads=4)
}
