% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovafilter.R
\name{anovafilter}
\alias{anovafilter}
\title{anovafilter:
filters a genedata via}
\usage{
anovafilter(dataset, nthreads = 2)
}
\arguments{
\item{dataset}{A transcriptomics dataset. First columns should be gene names. All other columns should be expression levels.}

\item{nthreads}{Number of processor threads for the filtering. Defaults to 2.}
}
\description{
anovafilter:
filters a genedata via
}
\examples{
Laurasmappings_filtered <- anovafilter(Laurasmappings, nthreads=4)
}
