% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{AnovaFilter}
\alias{AnovaFilter}
\title{AnovaFilter}
\usage{
AnovaFilter(dataset, threshold = 0.05, nthreads = NULL)
}
\arguments{
\item{dataset}{A transcriptomics dataset. First columns should be gene names. All other columns should be expression levels.}

\item{threshold}{Set the p-value threshold for the filtering}

\item{nthreads}{Number of processor threads for the filtering. If not specifed then the maximum number of logical cores are used.}
}
\description{
Filters a gene activity dataframe via ANOVA
}
\examples{
Laurasmappings_filtered <- anovafilter(Laurasmappings, nthreads=4)

}
