% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircadianTools.R
\docType{package}
\name{CircadianTools}
\alias{CircadianTools}
\alias{CircadianTools-package}
\title{CircadianTools: A Collection of Tools for Detecting Rhythmic Genes}
\description{
Allows analysis of rhythmic genes to be easily carried out on transcriptomics data using R. Designed to be as flexible as possible such as by allowing an unequal number of replicates. \cr \cr
}
\section{Simple Plots}{

\code{\link{BasicPlot}}: Plots activity data as points and average activity as lines \cr \cr
\code{\link{CompPlot}}: Plots two genes from a gene activity dataset. \cr \cr
\code{\link{DatasetPlot}}: Saves plots of all genes in a dataset. \cr \cr
}

\section{Clustering}{

\code{\link{ClusterOverview}}: Plots the mean and error bars for all clusters across time. \cr \cr
\code{\link{Clusterplot}}: Plots the mean and error bars for the genes in a cluster across time. \cr \cr
\code{\link{Clusterspread}}: Shows how many genes are in each cluster after clustering has been applied. \cr \cr
\code{\link{ClustersText}}: Takes a dataframe of clusters and stores the name of all genes in a text file. The row number deontes the cluster number. \cr \cr
\code{\link{ClusterTimeProfile}}: Provides a dataframe of median values at each time point for each cluster. \cr \cr
\code{\link{HClustering}}: Applies hierarchical clustering, clustering to a transcriptomics dataset and appends a cluster column to this dataset for all genes. \cr \cr
\code{\link{PamClustering}}: Applies PAM (Partitioning around Medoids) clustering to a transcriptomics dataset and appends a cluster column to this dataset for all genes. \cr \cr
}

\section{Correlation}{

\code{\link{CorAnalysis}}: Ranks correlation between a given gene and all other genes in a dataset. Plots both the given gene and highly correlated genes for a given correlation value. \cr \cr
\code{\link{CorAnalysisCluster}}: Correlates the average activity of a cluster with the average activity of every other cluster. \cr \cr
\code{\link{CorAnalysisDataset}}: Correlates every gene in a dataset with every other gene in the same dataset. Allows a timelag between genes to be correlated. \cr \cr
\code{\link{CorAnalysisClusterDataset}}: Correlates the average activity of each cluster with every other cluster in a dataset. \cr \cr
\code{\link{CorAnalysisPar}}: Parallel Implementation of \code{CorAnalysis}. \cr \cr
\code{\link{CorSignificantPlot}}: Prints or saves the genes found to be most significant by \code{CorAnalysis}. \cr \cr
}

\section{Cosinor Functions}{

\code{\link{CosinorAnalysis}}: Fits cosinor models to transcriptomics data and plots the best-fitting models using ggplot2. \cr \cr
\code{\link{CosinorAnalysisPar}}: Parallel Implementation of \code{CosinorAnalysis}. \cr \cr
\code{\link{CosinorPlot}}: Fits a cosinor model to a given gene in a given dataset and plots the model. \cr \cr
\code{\link{CosinorSignificantPlot}}: Prints or saves the genes found to be most significant by \code{cosinoranalysis}. \cr \cr
}

\section{Filtering}{

\code{\link{AnovaFilter}}: Filters a gene activity dataframe via ANOVA. \cr \cr
\code{\link{CombiFilter}}: Filters a transcriptomics dataset by using \code{Zerofilter}, \code{Anovafilter} and \code{Sizefilter}. \cr \cr
\code{\link{SizeFilter}}: Filters the genes with the smallest range from a transcriptomics dataset. \cr \cr
\code{\link{TFilter}}: Experimental! Applies a filter where a t-test is carried out on gene activity levels between time points. \cr \cr
\code{\link{ZeroFilter}}: Filters a transcriptomics dataset such that there is a minimum number of non-zero activity readings for each gene. \cr \cr
}

\section{RAIN Functions}{

\code{\link{RainAnalysis}}: Carries out RAIN analysis on a gene dataset.  \cr \cr
\code{\link{RainSignificantPlot}}: Prints or saves the genes found to be most significant by \code{rainanalysis}. \cr \cr
}

\section{Turning Point Functions}{

\code{\link{TurningPlot}}: Fits a spline to a given gene in a given dataset. Finds the turning points. Plots the turning points and spline.\cr \cr
}

\section{Utility Functions}{

\code{\link{CytoscapeFile}}: Converts a correlation dataframe object into a format suitable for cytoscape and saves as a csv file. \cr \cr
\code{\link{FileConflict}}: Checks if a file which will be created already exists and, if necessary asks the user if this file should be overwritten. \cr \cr
\code{\link{GeneClean}}: Removes columns and rows which show no gene activity over time. \cr \cr
\code{\link{GeneRange}}: Finds the range of gene activity for each gene in a dataframe. The median for the replicates is used for each time point. \cr \cr
\code{\link{GeneScale}}: Centers/scales every gene in a transcriptomics dataset. \cr \cr
\code{\link{GeneSub}}: Takes an object where the first column is genenames (IE a column of known Circadian genes) and subsets from a dataset containing activity for these genes. \cr \cr
\code{\link{ggplot.cosinor.lm}}: Adapted from the Cosinor package by Michael Sachs. Given a cosinor.lm model fit, generate a plot of the data with the fitted values.\cr \cr
\code{\link{MakeTimevector}}: Produces a vector of time values for the gene activity readings. \cr \cr
\code{\link{MedList}}: Provides a dataframe of median values at each time point for each gene from a transcriptomics dataset. \cr \cr
\code{\link{TAnalysis}}: Experimental! A t.test is carried out on gene activity levels between time points and the number of significant increases & decreases is returned. \cr \cr
}

