% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_param_selection.R
\name{ClusterParamSelection}
\alias{ClusterParamSelection}
\title{ClusterParamSelection}
\usage{
ClusterParamSelection(dataset, k = c(2, 5, 10), method = c("pam",
  "hclust", "diana"), metric = "euclidean", nthreads = 2,
  save.plot = TRUE, save.df = TRUE, path = NULL)
}
\arguments{
\item{dataset}{A transcriptomics dataset. Preferably filtered first. First columns should be gene names. All other columns should be expression levels.}

\item{k}{A numeric vector giving the number of clusters to be evaluated.}

\item{method}{The clustering method(s) to be used. Multiple methods can be considered by providing a character vector. Currently accepts 'pam', 'hclust' and 'diana'}

\item{metric}{The distance metric to be used to calculate the distances between genes. See parallelDist::parDist for all accepted arguments. Also allows the option of "abs.correlation".}

\item{nthreads}{The number of threads to be used for parallel computations.If NULL then the maximum number of threads available will be used.}

\item{save.plot}{Logical. If TRUE then saves the plots generated}

\item{save.df}{Logical. If TRUE then saves the validation metric results as a csv file.}

\item{path}{The directory to be used for saving plots and the validation metric results to. Uses the name of the dataset object appended with '_validation' if this argument is not specified}
}
\description{
Calculates validation metrics for different clustering methods and varying the number of partitions. The validation metrics are plotted.
}
\examples{
filter.df <- CombiFilter(Laurasmappings)
k.options <- seq(10,100, by=10)
hclust.validation <- ClusterParamSelection(filterdf, k=k.options)
}
