% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_functions.R
\name{CorAnalysis}
\alias{CorAnalysis}
\title{CorAnalysis:}
\usage{
CorAnalysis(genename, dataset, threshold = 0.9, average = "median",
  lag = 0, save = FALSE, print = TRUE, df = TRUE)
}
\arguments{
\item{genename}{the name of a gene intended for comparison with all other genes in the dataset. Must be a string.}

\item{dataset}{A transcriptomics dataset. First columns should be gene names. All other columns should be expression levels.}

\item{threshold}{Set correlation threshold value for which genes are considered significant and thus plotted. Defaults to 0.9}

\item{average}{The average to be used for comparing the time points. Either 'median' or 'mean'.}

\item{lag}{Setting any value other than 0 allows a gene to be correlated with lagged genes in the dataset. The number denotes the number of timesteps to lag by.}

\item{save}{Logical. If TRUE, saves plots to working directory. Defaults to FALSE.}

\item{print}{Logical. If TRUE renders highly correlated genes in the plot viewer. Defaults to TRUE}

\item{df}{Logical. If TRUE a dataframe containing the correlations of the given gene with all genes in the dataset is returned. Defaults to TRUE.}
}
\value{
Prints or saves ggplot2 object(s). Optionally returns dataframe containing gene names and correlation values
}
\description{
Ranks correlation between a given gene and all other genes in a dataset. Plots both the given gene and highly correlated genes for a given correlation value
}
\examples{
cor_results <- coranalysis(Laurasmappings, 'comp100002_c0_seq2')

}
