% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_functions.R
\name{CorAnalysisPar}
\alias{CorAnalysisPar}
\title{CorAnalysisPar:}
\usage{
CorAnalysisPar(genename, dataset, lag = 0, average = "median", nthreads = NULL)
}
\arguments{
\item{genename}{the name of a gene intended for comparison with all other
genes in the dataset. Must be a string.}

\item{dataset}{A transcriptomics dataset. First columns should be gene names.
All other columns should be expression levels.}

\item{lag}{Setting any value other than 0 allows a gene to be correlated with
lagged genes in the dataset. The number denotes the number of timesteps
to lag by.}

\item{average}{The average to be used for comparing the time points. Either
'median' or 'mean'.}

\item{nthreads}{Number of processor threads for the process. If not specifed
then the maximum number of logical cores are used.}
}
\value{
Returns dataframe containing gene names and correlation values
}
\description{
Parallel Implementation of \link{CorAnalysis}. Ranks correlation
 between a given gene and all over genes in a dataset. Plots both the given
 gene and highly correlated genes for a given correlation value
}
\examples{
cor_results <- CorAnalysisPar('comp100002_c0_seq2', Laurasmappings,
                              nthreads = 2)

}
