% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinor_functions.R
\name{CosinorSignificantPlot}
\alias{CosinorSignificantPlot}
\title{CosinorSignificantPlot :}
\usage{
CosinorSignificantPlot(results, dataset, number = 10, period = 24,
  print = TRUE, save = FALSE)
}
\arguments{
\item{results}{A dataframe generated by \code{CosinorAnalysis}}

\item{dataset}{A transcriptomics dataset which was used with \code{CosinorAnalysis} to generate the results dataframe. First columns should be gene names. All other columns should be expression levels.}

\item{number}{The number of most significant genes printed or saved}

\item{period}{The period of rhythmicity which is being tested for. Defaults to 24 (circadian).}

\item{print}{Logical. If TRUE renders significant genes in the plot viewer. Defaults to TRUE}

\item{save}{Logical. If TRUE, saves plots to working directory. Defaults to FALSE.}
}
\value{
Prints or saves ggplot2 object(s).
}
\description{
Prints or saves the genes found to be most significant by \link{CosinorAnalysis}
}
\examples{
cosinor_results <- CosinorAnalysis(Laurasmappings)
CosinorSignificantPlot(cosinor_results, Laurasmappings, number = 15, period=24 ,save=TRUE)

}
