% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoscapefile.R
\name{CytoscapeFile}
\alias{CytoscapeFile}
\title{CytoscapeFile:}
\usage{
CytoscapeFile(cor.dataset, filename = NULL, nthreads = NULL)
}
\arguments{
\item{cor.dataset}{A NxN datframe of correlation values created by \link{coranalysisdataset} or \link{coranalysisclusterdataset}}

\item{nthreads}{Number of processor threads for the process. If not specifed then the maximum number of logical cores are used.}
}
\value{
Dataframe object in the new format.
}
\description{
Converts a correlation dataframe object into a format suitable for cytoscape and saves as a csv file.
}
