% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoscape_functions.R
\name{CytoscapeFilter}
\alias{CytoscapeFilter}
\title{CytoscapeFilter}
\usage{
CytoscapeFilter(cytoscape.file, threshold = 0.9, save = TRUE,
  filename = NULL)
}
\arguments{
\item{cytoscape.file}{A dataframe object generated by CytoscapeFile}

\item{threshold}{Positive genes below this threshold and negative genes above
the additive inverse of this threshold will be removed.}
}
\description{
Reduces the size of a file intended for Cytoscape by filtering
 out the genes/clusters which are not correlated
}
\examples{
filter.df <- CombiFilter(Laurasmappings)
pam.df <- PamClustering(filter.df)
cor.df <- CorAnalysisClusterDataset(pam.df, save=FALSE)
cyto.df <- CytoscapeFile(cor.df, save=FALSE)
cyto.filtered.df <- CytoscapeFilter(cyto.df, threshold=0.95)

}
