% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{FindClusterQuantile}
\alias{FindClusterQuantile}
\title{FindClusterQuantile}
\usage{
FindClusterQuantile(cluster.no, centers.df, metric = "euclidean")
}
\arguments{
\item{cluster.no}{The number which identifies the cluster.}

\item{centers.df}{Centers of clusters generated by
\link{ClusterCenterGenerator}.}

\item{metric}{The distance metric to be used to calculate the distances
between genes. See parallelDist::parDist for all accepted arguments. Also
allows the option of 'abs.correlation'.}
}
\description{
Finds The distances between the center of a cluster and the
centers of all other clusters.
}
\examples{
pam.df <- PamClustering(Laurasmappings, k = 10, nthreads = 2)
cluster.centers <- ClusterCenterGenerator(pam.df, nthreads = 2)
FindClusterQuantile(1, cluster.centers)
}
