% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_distances.R
\name{FindClusterQuantile}
\alias{FindClusterQuantile}
\title{FindClusterQuantile}
\usage{
FindClusterQuantile(cluster.no, centers.df)
}
\arguments{
\item{cluster.no}{The number which identifies the cluster}

\item{centers.df}{Centers of clusters generated by
\link{ClusterCenterGenerator}}
}
\description{
Finds The distances between the center of a cluster and the
centers of all other clusters.
}
\examples{
filter.df <- CombiFilter(Laurasmappings)
pam.df <- PamClustering(filter.df, k = 75)
cluster.centers <- ClusterCenterGenerator(cluster.dataset, nthreads = nthreads)
FindClusterQuantile(cluster.centers)
}
