% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneclean.R
\name{GeneClean}
\alias{GeneClean}
\title{GeneClean:}
\usage{
GeneClean(dataset)
}
\arguments{
\item{dataset}{A transcriptomics dataset. First columns should be gene names. All other columns should be expression levels.}
}
\value{
A 'cleaned' transcriptomics dataset
}
\description{
Removes columns and rows which show no gene activity over time
}
\examples{
cleaned.df<-GeneClean(Laurasmappings)

}
