% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_param_selection.R
\name{HclustParamSelection}
\alias{HclustParamSelection}
\title{HclustParamSelection}
\usage{
HclustParamSelection(dataset, k = c(2, 5, 10), nthreads = 4)
}
\arguments{
\item{dataset}{A transcriptomics dataset. Preferably filtered first. First columns should be gene names. All other columns should be expression levels.}

\item{k}{A numeric vector giving the number of clusters to be evaluated.}

\item{nthreads}{The number of threads to be used for parallel computations.If NULL then the maximum number of threads available will be used.
filter.df <- CombiFilter(Laurasmappings)
k.options <- seq(10,100, by=10)
hclust.validation <- hclustParamSelection(filterdf, k=k.options)}
}
\description{
Runs hierarchical clustering with differing numbers of partitions and returns validation metrics.
}
