% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_param_selection.R
\name{HclustParamSelection}
\alias{HclustParamSelection}
\title{HclustParamSelection}
\usage{
HclustParamSelection(dataset = NULL, distance = NULL, k = c(2, 5,
  10), scale = TRUE, metric = "euclidean", nthreads = 4)
}
\arguments{
\item{dataset}{A transcriptomics dataset. Preferably filtered first. First columns should be gene names. All other columns should be expression levels.  Not needed if an argument to distance is given.}

\item{distance}{A distance matrix. If a distance matrix has already been created (such as by using the DistanceGen function), the matrix can be passed to this function to save time. If a distance matrix is not provided then it will be generated by the function.}

\item{k}{A numeric vector giving the number of clusters to be evaluated.}

\item{scale}{Logical. If TRUE then each gene will be scaled}

\item{metric}{The distance metric to be used to calculate the distances between genes. See parallelDist::parDist for all accepted arguments. Also allows the option of 'abs.correlation'. Not used if a distance matrix is provided.}

\item{nthreads}{The number of threads to be used for parallel computations.If NULL then the maximum number of threads available will be used.
filter.df <- CombiFilter(Laurasmappings)
k.options <- seq(10,100, by=10)
hclust.validation <- hclustParamSelection(filterdf, k=k.options)}
}
\description{
Runs hierarchical clustering with differing numbers of partitions and returns validation metrics.
}
