% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{MedList}
\alias{MedList}
\title{MedList}
\usage{
MedList(dataset, nthreads = NULL)
}
\arguments{
\item{dataset}{A transcriptomics dataset. First columns should be gene names. All other columns should be expression levels.}

\item{nthreads}{Number of processor threads for the filtering. If not specifed then the maximum number of logical cores are used.}
}
\description{
Provides a dataframe of median values at each time point for each gene from a transcriptomics dataset.
}
\examples{
mediandf <- MedList(Laurasmappings, nthreads=4)

}
