% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{MedList}
\alias{MedList}
\title{MedList}
\usage{
MedList(dataset, nthreads = NULL)
}
\arguments{
\item{dataset}{A transcriptomics dataset. First columns should be gene names. All other columns should be expression levels.}
}
\description{
A Wrapper for MedList. Calls MedlistSeq if a singlethreaded function is required or MedListPar if a multithreaded function is needed.
}
\examples{
mediandf <- MedList(Laurasmappings, nthreads=4)

}
