% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medlist.R
\name{medlist}
\alias{medlist}
\title{medlist}
\usage{
medlist(dataset, nthreads = NULL)
}
\arguments{
\item{dataset}{A transcriptomics dataset. First columns should be gene names. All other columns should be expression levels.}

\item{nthreads}{Number of processor threads for the filtering. If not specifed then the maximum number of logical cores are used.}
}
\description{
Provides a dataframe of median values at each time point for each gene from a transcriptomics dataset.
}
\examples{
mediandf <- medlist(Laurasmappings, nthreads=4)

}
