% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pamclustering.R
\name{pamclustering}
\alias{pamclustering}
\title{pamclustering:}
\usage{
pamclustering(dataset, k, metric = "euclidean", nthreads = NULL,
  scale = FALSE, center = TRUE)
}
\arguments{
\item{dataset}{A transcriptomics dataset. First columns should be gene names. All other columns should be expression levels.}

\item{k}{The total number of clusters.}

\item{metric}{The distance metric to be used to calculate the distances between genes. See parallelDist::parDist for all accepted arguments.}

\item{nthreads}{Number of processor threads to be used for calculating the distance matrix. If not specifed then the maximum number of logical cores are used.}

\item{scale}{If the gene activity should be scaled before clustering.}

\item{center}{If the gene activity should be centered before clustering.}
}
\value{
Returns transcriptomics dataset provided with additional cluster column appended denoted which cluster each gene belongs to.
}
\description{
Applies PAM (Partitioning around Medoids) clustering to a transcriptomics dataset and appends a cluster column to this dataset for all genes.
}
\examples{
pam.df <- pamclustering(Laurasmappings,20)
}
