% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_param_selection.R
\name{PamParamSelection}
\alias{PamParamSelection}
\title{PamParamSelection}
\usage{
PamParamSelection(dataset = NULL, distance = NULL, k = c(2, 5, 10),
  metric = "euclidean", scale = TRUE, nthreads = 2)
}
\arguments{
\item{dataset}{A transcriptomics dataset. Preferably filtered first. First columns should be gene names. All other columns should be expression levels. Not needed if an argument to distance is given.}

\item{distance}{A distance matrix. If a distance matrix has already been created (such as by using the DistanceGen function), the matrix can be passed to this function to save time. If a distance matrix is not provided then it will be generated by the function.}

\item{k}{A numeric vector giving the number of clusters to be evaluated.}

\item{metric}{The distance metric to be used to calculate the distances between genes. See parallelDist::parDist for all accepted arguments. Also allows the option of 'abs.correlation'. Not used if a distance matrix is provided.}

\item{scale}{Logical. If TRUE then each gene will be scaled}

\item{nthreads}{The number of threads to be used for parallel computations. Warning! Increasing this value will massively increase RAM usage. If NULL then the maximum number of threads available will be used.}
}
\description{
Runs PAM with differing numbers of partitions and returns validation metrics.
}
\examples{
filter.df <- CombiFilter(Laurasmappings)
k.options <- seq(10,100, by=10)
pam.validation <- PamParamSelection(filterdf, k=k.options)
}
