% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainanalysis.R
\name{rainanalysis}
\alias{rainanalysis}
\title{rainanalysis :}
\usage{
rainanalysis(dataset, period)
}
\arguments{
\item{dataset}{A transcriptomics dataset. First columns should be gene names. All other columns should be expression levels.}
}
\value{
A dataframe object detailing the result of the rain analysis.
}
\description{
Carries out RAIN analysis on a gene dataset
}
\examples{
results <- rainanalysis(Laurasmappings)
}
