% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turningplot.R
\name{turningplot}
\alias{turningplot}
\title{Turningplot:
Fits a spline to a given gene in a given dataset. Finds the turning points. Plots the turning points and spline.}
\usage{
turningplot(genename, dataset, print = TRUE, save = FALSE)
}
\arguments{
\item{genename}{The name of a gene intended for plotting. Must be a string.}

\item{dataset}{A transcriptomics dataset. First columns should be gene names. All other columns should be expression levels.}

\item{print}{Logical. If TRUE renders plot in the plot viewer. Defaults to TRUE}

\item{save}{Logical. If TRUE, saves plot to working directory. Defaults to FALSE.}
}
\value{
Prints or saves ggplot2 object(s)
}
\description{
Turningplot:
Fits a spline to a given gene in a given dataset. Finds the turning points. Plots the turning points and spline.
}
\examples{
turningplot('comp100000_c0_seq2',LauraSingleMap)
}
